create PersistentTerminalProcess and move terminal and environmentVariable files from workbench -> platform
