+++ src/vs/platform/terminal/common/environmentVariable.ts
